Added pages. Added a chain feature for reports.
Added a reportChain column. Chains together edits based on the original
report that was filed. The page selection feature is a litle buggy due
to the fact that reports can be hidden from view and when I select 30 of
them based on what was hidden from view, I get some overlaps between
pages.