Changed how reports show on the main page. Chained reports also show up on the view pages. Edited some CSS.
Reports only show a few of the things on the main page now. Will
probably reduce them even more. Chained reports show up in a blob on the
view pages. Will fix them to just show the numbers. Probably will use
MySQL.Simple for that because I only need to select the ID's and not the
entire row.